+++ CHANGELOG.md
@@ -4,6 +4,24 @@ If you like this project and find it useful, please consider giving it a star on
 
 All notable changes to this project will be documented in this file.
 
+## [2.3.1] - 2024-12-12
+
+### Added
+
+- [colorControl] Update ColorControl cluster
+- [levelControl] Update currentLevel to minLevel
+
+### Changed
+
+- [package]: Updated dependencies.
+- [plugin]: Requires Matterbridge 1.6.6.
+
+### Fixed
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="./yellow-button.png" alt="Buy me a coffee" width="120">
+</a>
+
 ## [2.3.0] - 2024-12-04
 
 ### Added